Moved the Memory Manager's flags out of memoryStream.h and into a separate header, mmFlags.h.
This is the start of a more flexible, more pwoerful memory manager overall. I'm about to put work in to cater for all kinds of desperate memory management situations.

Along with the next set of changes will be a radical change in the way that the memory manager allocates, both in the Memory stream and the NUMA Tributary.

When this next batch of design changes is done, the scheduler will be my next big aim.