In light of the phase of analysis I'm going through with each inidividual subsystem of the kernel now, I'm going to have to re-design a lot of things in the memory manager.
This commit changes the way that memory ranges are added to a NUMA Memory Bank. I've replaced the old array with a linked list, sicne this will allow for
better concurrency protection of the member range pointers.

And now, the __kspacePtrMem has to be amended to be at least the size of two
pointers.