Began work on libzbzcore; Created paths for Distributary and Driver processes. Streamlined the kernel's entry points for different exec formats and process types.
The kernel now proceeds up to __korientationMain4().

To deal with the issue of continuity of access to the response message for a
newly spawned process, we now store a pointer to the process' response message
in its PCB. This message can only be sent once. This enables libzbzcore in
userspace to properly syscall into the kernel and notify the creating process
of the status of the child process.