Pretty beautiful patch, if you ask me, or really anyone else who has worked with this source (nobody else). I've managed to create a really nice continuation and callback framework that makes callbacks and async very clean, though somewhat taciturn.
Callback objects can be used as the "privateData" for syscalls. These Callback
objects have polymorphic methods that will correctly call the callback function
they contain. This enables us to use completely callback-agnostic event loops,
and get type-safe callbacks for all async calls.

The only real drawback is that this new approach mandates memory allocation for
every async call. We cannot reasonably find a way to hardcode the Callback class
instances within MessageStream::sHeader.

I was pleasantly surprised to find that the kernel executed to completion
almost immediately, and without a hitch. The new abstraction is *very* clean in
that sense -- it was integrated very quickly and very painlessly.

This new callback framework removes the need for MessageStream::sIterator.
Also makes message passing (in theory) as quick as moving a pointer around.

I hope for the next 2-3 patches to introduce the cleanup for the per-cpu thread
fiasco.

* Callbacks:
  * Added callback.h:
    * class Callback, a polymorphic base class (abstract) with an operator()
      that can be invoked to execute the callback function contained inside of
      the object.
    * __kCallback: Extrapolated class for very "standard" callbacks which may
      be common throughout the kernel.
  * Integrated Callback throughout the kernel except in maybe ZUI Server.
* Namespacing:
  * Namespaced all the monstrous code inside of __klibzbzcore.
  * Cleaned up all of that code as a result; it is much cleaner now, and far
    more easily readable and navigable. And that with a grain of salt, too.
* New stub for taskTrib.kill(). Wraps around taskTrib.dormant.