First step toward removing per-cpu threads. Replacing them with power threads. Each CPU now has an embedded thread that is its own and has the ID of its owning CPU. Removed former methods like getTaskContext(), etc.
Restructured the inheritance chain for Task, Thread and TaskContext. Thread
now directly inherits from Task and TaskContext. Task and TaskContext cannot
be immediately instantiated.

I am a bit taken aback (actually, I'm completely shocked) that the kernel ran
to completion on the first try after I got it building again. Naturally, I am
currently in a state of high alert as a result of that...