This patch is truly a very diverse patch in terms of focus.
I've added and improved iterators in some of the kernel's list classes. The ZUI
server thread is now a thread and nothing more; it is no longer a distributary
and the remnants of that form of it will be removed over time. We no longer
allow processes to direct their syscall responses to other threads, or other
processes. This feature has been removed from those APIs which had already
implemented it.

Improved the separation of function between the components of Floodplainn. Now
we have 3 logically distinct components: Floodplainn, ZUI and ZUDI. ZUDI will
soon be renamed to ZUE.

New driver: zrootdev. This is actually not a proper UDI driver, but an
embodiment of the kernel within a driver. The root device in the kernel device
tree is now "@f/by-id" itself.

* __kclasses:
  * HardwareIdList:
    * Improved Iterator.
      ^ It now works such that begin() returns an iterator that references the
        first valid object, and the "cursor" member has the correct index filled
        in. Previously, the cursor member would have the index of the next
        member, Y, while holding the value of a member, X.
      ^ Added all necessary operators to Iterator.
    * Renamed getLoopItem() to getNextItem().
    * Removed prepareForLoop(), and made getNextItem() private.
* API & functionality:
  * Kernel no longer supports response redirection for syscalls.
    ^ As a result, determineTargetThreadId() now always returns the ID of the
      target thread as passed to it.
* Added udi_bridge and udi_physio related headers to the kernel tree.
* UDI:
  * New driver: zrootdev.
  * Kernel root device is now @f/by-id.
  * Did a whole lot of tedious namespacing of the Floodplainn, ZUI and ZUDI
    components.
  * Now the default action for new device indication is set by default in a
    global variable. No longer need to syscall to tell it what to do.
  * Each new component also has its own MessageStream queue.
* Prototyped FloodplainnStream.
* Renamed floodplainn_zuiServer_*End.cpp to zui*End.cpp.
* Added new kernel-nomenclature device classes.

* Kernel now enumerates @f/by-id in fplainn::Zui::initialize().
  ^ We now allow getDevice() to return handles to @f/by-id.

* Fixed a minor bug in fvfs::Currentt::getPath(), where stripping away @f from
  a path after stripping away slashes may not have worked.