We now have the basic wireframing classes for building the UDI channel abstraction and integrating it into the kernel for the interconnection of both drivers and application processes.
So far spawnChildBindChannel() is untested, but spawnInternalBindChannel() is
tested (not very rigorously though). We make very heavy use of both polymorphism
and multiple inheritance in the channel manipulation code, so take note.

But all in all, very exciting patch since it means that within a few more
patches, we'll be "blasting" messages about, as someone on the LKML described it
some years back.

* createChannel() functions:
  * Cleaned these up by making them take an Incomplete*Channel pointer, instead
    of passing 6+ arguments on the stack.
  * They are now static methods of their logical container entities.

* Destroyed the older Channel classes which were defined as subclasses of
  DeviceInstance, and moved the entire hierarchy along with class Region out
  into the fplainn:: namespace.
* Prototyped the FloodplainnStream:: namespace channel spawning and control
  APIs.
* Moved some functions around in source files, mostly between
  udiServiceCalls.cpp and zudi.cpp to ensure sane namespacing.