This patch. THIS patch. This PARTICULAR patch. Eh, boi! When yuh big, yuh large.
This patch introduces all the code necessary to package and marshal a UDI call
into a kernel message and send it across a channel, from endpoint to endpoint.
It can successfully transfer a call across a D2D or a D2S channel. Only
D2S channels have been tested for now.

* Implemented fplainn::Zudi::send().
* Implemented FloodplainnStream::send().

	^ These two function both take a kernel endpoint and marshal a UDI
	  call over it. They first validate the endpoint pointers passed to
	  them, then delegate most of the work over to
	  fplainn::sChannelMsg::send() to actually marshal the call into a
	  message, and eventually post it to the target thread.

* Re-wrote large portions of udi_mei_call().
  ^ Now it properly gathers the layout data required to properly marshal the
    visible, marshal and DRIVER_TYPED portions of a call.

* Wrote some supporting methods that will help us to get proper inline member
  support later on.