This patch is pretty nice for me from a personal perspective since it represents my return to working regularly on Zambesii after several months of stochastic contribution.
It sees the introduction of a long-awaited ResizeableArray class, and the
beginning of the Zambesii UDI DMA framework implementation. We've dealt with
UDI DMA constraints in here, and skeleton-written the core classes for the
physical side of DMA.

The buffer and virtual mapping side of it will come later. Also, renamed the
files and class names of some of the kernel data structure classes. I also want
to take a look at the kernel data structure classes' iterators again:
Iterator::operator*(void) should return a REFERENCE to, and not a copy of the
underlying item.

Build system:
* The kernel should build without problems if downloaded from scratch from a
  repository.

__kclasses:
* Renamed PtrDblList to HeapDoubleList.
* Renamed PtrList to HeapList.
* Renamed SortedPtrDblList to SortedHeapDoubleList
* New ResizeableArray class, fully tested, very convenient.

General:
* Added a realloc() stand-in until we properly enable the heap to support
  realloc() -- its name is crudeRealloc().
  ^ Funny thing is that it's in <new> instead of <stdlib.h>.

UDI:
* DMA:
  * Wrote skeletons for classes:
    * MappedDmaScatterGatherList: contains both the virtual pages which map to
      the frames in a SGList, as well as the methods for transparently writing
      to the frames transparently.
    * ScatterGatherList: Represents a UDI SGList in its physical form.
    * DmaConstraints: Manages DMA Constraints setting and resetting, etc.
      ^ Has a really neat dump() function.
    * DmaRequest: This is the motherlode. This class manages the entire DMA
      transaction; from allocating the frames, to assembling the SGList, to
      mapping it for reading and writing, to eventually mapping a buffer to it.