This patch introduces UDI DMA Constraints parsing and compilation into kernel metadata as well as the foundation for implementing our constrained PMM allocation feature.
The constrained PMM allocation feature will be coming next, and it will
signal the beginning of the tail end of the kernel's DMA support.

After these features are complete, our kernel will finally be "real", in the
sense that it will be able to support the creation of systems running on top of
it that actually control the real world.

* Stylistic tabulation fixes.
* Fixed new compiler warnings.
* New preprocessor #define which expands to nothing, but is used to denote
  a declaration vs a definition.
* Notes on UDI's DMA design from a discussion with XgF in #udi on Freenode.

PMM:
* Constrained PMM prototype in both MemBmp and NumaMemoryBank.

UDI:
* Refined the DmaConstraints, DmaRequest and related classes' constructors.
* UDI DMA Constraints compiler class: DmaConstraints::Compiler.