Add support for IBM-PC RS232 and fix bugs in DmaConstraints::Compiler
This patch adds a working RS232 driver for the IBM-PC chipset, and adds it to
the debugPipe's device list, as well as enabling it as an output surface at
boot. The Rs232 code was pretty old, and it still used the Rivulet API.

It also fixes 2 bugs in the DmaConstraints::Compiler class.