MemBmp::constrainedGetFrames: Don't always allocate the max element size
We always attempted to allocate the max element size, even on the final iteration
of the allocation loop and this resulted in us wasting memory.

The reason for trying to allocate the max element size where possible was to reduce the
number of scgth elements required to describe the allocated memory.