Fplainn::Driver: We now track ops vectors
Tracking this information enables the kernel to know the name of the
metalanguage that enumerated a device. Specifically:

	udi_mgmt_enumerate_ack(..., ops_idx);

The ops_idx is used to tell the parents being bound to this new child what
ops vector to bind on and by extension, what metalanguage is being used
for binding.

Without this information we can't verify whether or not a parent-child pair
are being correctly bound together.