FplainnStream: add setChildConstraints and getParentConstraints
* Also add remapScatterGatherList.

SetChildCons enables a parent to constrain its child. Constraints are set on
the bind channel between parent and child. Children can read the
constraints set on them by their parent.

The reason it's important to allow a per-parent constraints mapping is because
a device may have different parents and buffers sent to different parents may
need to be constrained differently.