UDI: Generate meta_init_info caches from requirements, not metas
Previously, we generated meta_init_info related metalanguage metadata from
all 'meta' udiprops lines.

We saw that this doesn't make sense since there are derivative metalanguages
such as all the derivative metas made from udi_physio -- and these derivative
metalanguages do not have udi_meta_info structs.

So, we switch from dealing with 'meta' lines to 'provides' and 'requires' lines
in the udiprops files.

In the case of udi_mgmt, we may decide to make `requires udi` be the stand-in
for udi_mgmt.