__korientation,IRQs: Re-enable IRQs and local #INT on BSP CPU.
Context: In commits 8804624b9, b47235eff and 50b55d722 we began
rearranging the kernel boot to strip the kernel down to a microkernel.
We decided to no longer support watchdog petting __using timers__
at boot because and instead, we decided to require that some other
chipset/firmware module handles or sets up the watchdog responsibly.

NEW DECISION:
We can just pet the watchdog regularly in software during the boot
until we have timers up. This doesn't require a regular timer signal.
We can just pet it at arbitrary milestones during the bootup
sequence.

So we can now continue to strip the kernel down to be more like a
microkernel.