RecursiveLock: Use CMPXCHG, eliminate taskId waitlock
We can entirely eliminate the taskId lock by using CMPXCHG. It allows
us to atomically swap in a contender thread's TID if the lock is free,
and only if the lock is free. This completely eliminates the need for
the taskId lock. We can now just directly use Lock::lock.