CpuTrib: Don't panic on cpuId >= MAX_NCPUs, unless it's BSP
AP CPUs whose cpuId >= MAX_NCPUs are simply ignored and a warning
is printed to the klog.

If the BSP CPU's cpuId >= MAX_NCPUs, that's a problem because it means
we have no powerstack for it, and this could mean we won't be able
to wake it back up if it gets powered down.