Debug: WIP. See current-progress.txt for a summary of where we were
See current-progress.txt for a summary of what our goals were. Basically
we were tyring to fix up the scheduler. We also reworked the way
the service threads in the kernel booted up, and added async bridging
mechanisms.

Sadly, we ended up not being able to get everything working in decent
time, and we encountered deadlocks; locks being held on kernel entry
(see --enable-debug-lock-exceptions) and failure to wakeup the
timerTrib eventProcessor thread in order for it to finish booting
AP CPUs.

Hence, the bootup process is halted at that bug (eventProcessor not
getting unblock()ed), and hence, AP CPUs not being booted.