IbmPc: Move .init_array above .init
.init is a substring of .init_array, so it was causing the
.init_array sections to get incorporaed into .init. This most
likely is the cause of the confusion we had with the alleged "ghost"
code in GCC that was placing constructors in a .init section when
we couldn't find any place in the code that was doing it.

It turns out that the i686-linux-gnu xtoolchain was placing the
constructors into .init_array, but were putting .init_array into
.init.

GCC never created a .init section at all.