MRLock,WaitLock: Allow local IRQ flag ownership transfer
This enables us to implement scenarios like the new ones we've come
across inside of the enqueueing and dequeueing operations inside of
SingleWaiterQueue and MessageStream.

Basically, if we nest lock operations in a non-mirrored way, then
the lock that releases() first may re-enable IRQs while the second
lock is still held. This will violate the critical section that should
be protected by the 2nd lock.

Hence, we now allow for IRQ flag ownership transfer from the 1st lock
to the 2nd lock.